+++ discord/commands/core.py
@@ -226,7 +226,7 @@ class ApplicationCommand(_BaseCommand, Generic[CogT, P, T]):
             "__default_member_permissions__",
             kwargs.get("default_member_permissions", None),
         )
-        self.nsfw: bool | None = getattr(func, "__nsfw__", kwargs.get("nsfw", None))
+        self.nsfw: bool | None = getattr(func, "__nsfw__", kwargs.get("nsfw", False))
 
         integration_types = getattr(func, "__integration_types__", kwargs.get("integration_types", None))
         contexts = getattr(func, "__contexts__", kwargs.get("contexts", None))
@@ -1182,7 +1182,7 @@ class SlashCommandGroup(ApplicationCommand):
 
         # Permissions
         self.default_member_permissions: Permissions | None = kwargs.get("default_member_permissions", None)
-        self.nsfw: bool | None = kwargs.get("nsfw", None)
+        self.nsfw: bool | None = kwargs.get("nsfw", False)
 
         integration_types = kwargs.get("integration_types", None)
         contexts = kwargs.get("contexts", None)